refactored memory store
